Remove previous implementations and rename the new ones
