--- core/retries/src/main/java/software/amazon/awssdk/retries/internal/DefaultStandardRetryStrategy2.java
@@ -1,91 +1,0 @@
-/*
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License.
- * A copy of the License is located at
- *
- *  http://aws.amazon.com/apache2.0
- *
- * or in the "license" file accompanying this file. This file is distributed
- * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
-
-package software.amazon.awssdk.retries.internal;
-
-import java.util.function.Predicate;
-import software.amazon.awssdk.annotations.SdkInternalApi;
-import software.amazon.awssdk.retries.StandardRetryStrategy;
-import software.amazon.awssdk.retries.api.BackoffStrategy;
-import software.amazon.awssdk.retries.internal.circuitbreaker.TokenBucketStore;
-import software.amazon.awssdk.utils.Logger;
-
-@SdkInternalApi
-public final class DefaultStandardRetryStrategy2
-    extends BaseRetryStrategy<StandardRetryStrategy.Builder, StandardRetryStrategy> implements StandardRetryStrategy {
-    private static final Logger LOG = Logger.loggerFor(DefaultStandardRetryStrategy2.class);
-
-    DefaultStandardRetryStrategy2(Builder builder) {
-        super(LOG, builder);
-    }
-
-    @Override
-    public Builder toBuilder() {
-        return new Builder(this);
-    }
-
-    public static Builder builder() {
-        return new Builder();
-    }
-
-    public static class Builder extends BaseRetryStrategy.Builder implements StandardRetryStrategy.Builder {
-
-        Builder() {
-        }
-
-        Builder(DefaultStandardRetryStrategy2 strategy) {
-            super(strategy);
-        }
-
-        @Override
-        public Builder retryOnException(Predicate<Throwable> shouldRetry) {
-            setRetryOnException(shouldRetry);
-            return this;
-        }
-
-        @Override
-        public Builder maxAttempts(int maxAttempts) {
-            setMaxAttempts(maxAttempts);
-            return this;
-        }
-
-        @Override
-        public Builder backoffStrategy(BackoffStrategy backoffStrategy) {
-            setBackoffStrategy(backoffStrategy);
-            return this;
-        }
-
-        @Override
-        public Builder circuitBreakerEnabled(Boolean circuitBreakerEnabled) {
-            setCircuitBreakerEnabled(circuitBreakerEnabled);
-            return this;
-        }
-
-        public Builder tokenBucketExceptionCost(int exceptionCost) {
-            setTokenBucketExceptionCost(exceptionCost);
-            return this;
-        }
-
-        public Builder tokenBucketStore(TokenBucketStore tokenBucketStore) {
-            setTokenBucketStore(tokenBucketStore);
-            return this;
-        }
-
-        @Override
-        public StandardRetryStrategy build() {
-            return new DefaultStandardRetryStrategy2(this);
-        }
-    }
-}
