+++ core/retries/src/test/java/software/amazon/awssdk/retries/internal/RetryStrategyCommonTest.java
@@ -273,7 +273,7 @@ class RetryStrategyCommonTest {
 
         @Override
         public TestCase configureTokenBucketMaxCapacity(int maxCapacity) {
-            ((DefaultStandardRetryStrategy2.Builder) builder).tokenBucketStore(
+            ((DefaultStandardRetryStrategy.Builder) builder).tokenBucketStore(
                 TokenBucketStore
                     .builder()
                     .tokenBucketMaxCapacity(maxCapacity)
@@ -283,13 +283,13 @@ class RetryStrategyCommonTest {
 
         @Override
         public TestCase configureTokenBucketExceptionCost(int exceptionCost) {
-            ((DefaultStandardRetryStrategy2.Builder) builder).tokenBucketExceptionCost(exceptionCost);
+            ((DefaultStandardRetryStrategy.Builder) builder).tokenBucketExceptionCost(exceptionCost);
             return this;
         }
 
         @Override
         public TestCase configureCircuitBreakerEnabled(boolean enabled) {
-            ((DefaultStandardRetryStrategy2.Builder) builder).circuitBreakerEnabled(enabled);
+            ((DefaultStandardRetryStrategy.Builder) builder).circuitBreakerEnabled(enabled);
             return this;
         }
     }
@@ -303,7 +303,7 @@ class RetryStrategyCommonTest {
 
         @Override
         public TestCase configureTokenBucketMaxCapacity(int maxCapacity) {
-            ((DefaultLegacyRetryStrategy2.Builder) builder).tokenBucketStore(
+            ((DefaultLegacyRetryStrategy.Builder) builder).tokenBucketStore(
                 TokenBucketStore
                     .builder()
                     .tokenBucketMaxCapacity(maxCapacity)
@@ -313,28 +313,28 @@ class RetryStrategyCommonTest {
 
         @Override
         public TestCase configureTokenBucketExceptionCost(int exceptionCost) {
-            ((DefaultLegacyRetryStrategy2.Builder) builder).tokenBucketExceptionCost(exceptionCost);
+            ((DefaultLegacyRetryStrategy.Builder) builder).tokenBucketExceptionCost(exceptionCost);
             return this;
         }
 
         @Override
         public TestCase configureCircuitBreakerEnabled(boolean enabled) {
-            ((DefaultLegacyRetryStrategy2.Builder) builder).circuitBreakerEnabled(enabled);
+            ((DefaultLegacyRetryStrategy.Builder) builder).circuitBreakerEnabled(enabled);
             return this;
         }
 
         public TestCaseForLegacy configureTreatAsThrottling(Predicate<Throwable> isThrottling) {
-            ((DefaultLegacyRetryStrategy2.Builder) builder).treatAsThrottling(isThrottling);
+            ((DefaultLegacyRetryStrategy.Builder) builder).treatAsThrottling(isThrottling);
             return this;
         }
 
         public TestCaseForLegacy configureThrottlingBackoffStrategy(BackoffStrategy backoffStrategy) {
-            ((DefaultLegacyRetryStrategy2.Builder) builder).throttlingBackoffStrategy(backoffStrategy);
+            ((DefaultLegacyRetryStrategy.Builder) builder).throttlingBackoffStrategy(backoffStrategy);
             return this;
         }
 
         public TestCaseForLegacy configureBackoffStrategy(BackoffStrategy backoffStrategy) {
-            ((DefaultLegacyRetryStrategy2.Builder) builder).backoffStrategy(backoffStrategy);
+            ((DefaultLegacyRetryStrategy.Builder) builder).backoffStrategy(backoffStrategy);
             return this;
         }
     }
@@ -349,7 +349,7 @@ class RetryStrategyCommonTest {
 
         @Override
         public TestCase configureTokenBucketMaxCapacity(int maxCapacity) {
-            ((DefaultAdaptiveRetryStrategy2.Builder) builder).tokenBucketStore(
+            ((DefaultAdaptiveRetryStrategy.Builder) builder).tokenBucketStore(
                 TokenBucketStore
                     .builder()
                     .tokenBucketMaxCapacity(maxCapacity)
@@ -359,18 +359,18 @@ class RetryStrategyCommonTest {
 
         @Override
         public TestCase configureTokenBucketExceptionCost(int exceptionCost) {
-            ((DefaultAdaptiveRetryStrategy2.Builder) builder).tokenBucketExceptionCost(exceptionCost);
+            ((DefaultAdaptiveRetryStrategy.Builder) builder).tokenBucketExceptionCost(exceptionCost);
             return this;
         }
 
         @Override
         public TestCase configureCircuitBreakerEnabled(boolean enabled) {
-            ((DefaultAdaptiveRetryStrategy2.Builder) builder).circuitBreakerEnabled(enabled);
+            ((DefaultAdaptiveRetryStrategy.Builder) builder).circuitBreakerEnabled(enabled);
             return this;
         }
 
         public TestCase configureTreatAsThrottling(Predicate<Throwable> isThrottling) {
-            ((DefaultAdaptiveRetryStrategy2.Builder) builder).treatAsThrottling(isThrottling);
+            ((DefaultAdaptiveRetryStrategy.Builder) builder).treatAsThrottling(isThrottling);
             return this;
         }
     }
